// benchmarks/solar-system/runs/fable/frontier-v2/src/data/types.ts
export type BodyId =
  | 'sun'
  | 'mercury' | 'venus' | 'earth' | 'mars' | 'jupiter' | 'saturn' | 'uranus' | 'neptune'
  | 'moon'
  | 'phobos' | 'deimos'
  | 'io' | 'europa' | 'ganymede' | 'callisto'
  | 'titan' | 'enceladus'
  | 'miranda' | 'ariel' | 'umbriel' | 'titania' | 'oberon'
  | 'triton'
  | 'pluto' | 'charon'
  | 'ceres' | 'eris' | 'haumea' | 'makemake'
  | 'vesta'
  | 'comet-observatory'
  | 'asteroid-belt' | 'kuiper-belt' | 'oort-cloud';
// ...
export interface LocalizedText {
  pt: string;
  en: string;
}